classifier/requirements.txt
fastapi==0.114.0
uvicorn==0.29.0
pydantic==2.6.3
torch==2.2.1
transformers==4.36.2
sentence-transformers==2.5.1
scikit-learn==1.4.2
pyyaml==6.0.1
prometheus-client==0.12.0
starlette-prometheus==0.9.0
gunicorn==21.2.0
numpy==1.26.4
joblib==1.3.2
sentencepiece==0.2.0
nltk==3.8.1

# Additional package indexes
--find-links https://download.pytorch.org/whl/torch_stable.html
--extra-index-url https://pypi.org/simple/ 